refactor: resolve composer defaults server-side
