+++ apps/app/src/hooks/cache-owners/mutation-cache-effects.ts
@@ -1,7 +1,6 @@
 import {
   projectPathsQueryKeyPrefix,
   sidebarNavigationQueryKey,
-  threadDefaultExecutionOptionsQueryKey,
   threadQueuedMessagesQueryKey,
   threadPromptHistoryQueryKey,
   threadQueryKey,
@@ -12,6 +11,7 @@ import {
   threadTimelineQueryKeyPrefix,
   threadTimelineTurnSummaryDetailsQueryKeyPrefix,
 } from "../queries/query-keys";
+import { threadDefaultExecutionOptionsQueryKey } from "../queries/thread-default-execution-options-query";
 import type {
   ProjectArg,
   QueryClientArg,
