+++ apps/app/src/hooks/useThreadCreationOptions.ts
@@ -1,4 +1,3 @@
-import { useAtom } from "jotai";
 import {
   type ComponentType,
   useCallback,
@@ -16,30 +15,39 @@ import type {
 } from "@bb/domain";
 import type {
   CreateExecutionInputSources,
-  ExecutionInputFieldSource,
   ExistingThreadExecutionInputSources,
   SystemExecutionOptionsModelLoadError,
 } from "@bb/server-contract";
 import { parseEnvironmentValue } from "@/components/pickers/environment-picker-value";
-import {
-  createLocalStorageEnumStorage,
-  createProjectScopedStorageAtomFamily,
-  rawStringLocalStorage,
-} from "@/lib/browser-storage";
 import { useRootComposeReuseEnvironment } from "@/lib/root-compose-selection";
 import { getProviderIconInfo } from "@/lib/provider-icon";
 import { reconcileReasoningLevel } from "@bb/domain";
 import { useSystemExecutionOptions } from "./queries/system-queries";
+import {
+  usePersistedEnvironmentSelection,
+  usePersistedModelSelection,
+  usePersistedPermissionModeSelection,
+  usePersistedProviderSelection,
+  usePersistedReasoningLevelSelection,
+  usePersistedServiceTierSelection,
+} from "./thread-creation-options/persisted-selection-fields";
+import {
+  buildExecutionInputSources,
+  formatModelLabel,
+  getInitialThreadPromptSelections,
+  resolvePermissionModeSelection,
+  syncUntouchedThreadPromptSelections,
+  type ScopedExecutionInputSources,
+  type ThreadPromptField,
+  type ThreadPromptSelections,
+  type UseComponentLocalCreationOptions,
+  type UseNewThreadCreationOptions,
+  type UsePromptModelReasoningOptions,
+  updateThreadPromptSelections,
+} from "./thread-creation-options/selection-state";
 
-const MODEL_STORAGE_KEY = "bb.promptbox.model";
-const SERVICE_TIER_STORAGE_KEY = "bb.promptbox.service-tier";
-const REASONING_STORAGE_KEY = "bb.promptbox.reasoning";
-const PERMISSION_MODE_STORAGE_KEY = "bb.promptbox.permission-mode";
-const ENVIRONMENT_STORAGE_KEY = "bb.promptbox.environment";
-const PROVIDER_STORAGE_KEY = "bb.promptbox.provider";
-type StoredServiceTier = "" | ServiceTier;
-type StoredReasoningLevel = "" | ReasoningLevel;
-type StoredPermissionMode = "" | PermissionMode;
+export { formatModelLabel, resolvePermissionModeSelection };
+
 const EMPTY_PROVIDERS: ProviderInfo[] = [];
 
 const REASONING_LABELS: Record<ReasoningLevel, string> = {
@@ -77,37 +85,6 @@ interface PickerOption<T extends string> {
   icon?: ComponentType<{ className?: string }>;
 }
 
-type ThreadCreationOptionsScope =
-  | "new-thread"
-  | "component-local";
-
-interface UsePromptModelReasoningOptions {
-  enabled?: boolean;
-  environmentId?: string;
-  scope?: ThreadCreationOptionsScope;
-  projectId?: string | null;
-  resetKey?: string | number | null;
-  initialProviderId?: string;
-  initialModel?: string;
-  initialServiceTier?: ServiceTier;
-  initialReasoningLevel?: ReasoningLevel;
-  initialPermissionMode?: PermissionMode;
-  initialEnvironmentSelectionValue?: string;
-}
-
-interface UseNewThreadCreationOptions extends UsePromptModelReasoningOptions {
-  scope?: "new-thread";
-}
-
-interface UseComponentLocalCreationOptions
-  extends UsePromptModelReasoningOptions {
-  scope: "component-local";
-}
-
-type ScopedExecutionInputSources =
-  | CreateExecutionInputSources
-  | ExistingThreadExecutionInputSources;
-
 type StringSelectionSetter = (value: string) => void;
 type ServiceTierSelectionSetter = (value: ServiceTier | undefined) => void;
 type ReasoningLevelSelectionSetter = (value: ReasoningLevel) => void;
@@ -142,353 +119,7 @@ interface UseThreadCreationOptionsResult<TExecutionInputSources> {
   executionInputSources: TExecutionInputSources;
 }
 
-interface ThreadPromptSelections {
-  selectedProviderId: string;
-  selectedModel: string;
-  serviceTier: ServiceTier | undefined;
-  reasoningLevel: ReasoningLevel;
-  permissionMode: PermissionMode;
-  environmentSelectionValue: string;
-}
-
-type ThreadPromptField = keyof ThreadPromptSelections;
-
-interface ResolveCreateExecutionInputSourceArgs {
-  hasStoredValue: boolean;
-  hasValue: boolean;
-  touched: boolean;
-}
-
-interface StoredCreateExecutionValues {
-  selectedProviderId: string;
-  selectedModel: string;
-  serviceTier: StoredServiceTier;
-  reasoningLevel: StoredReasoningLevel;
-  permissionMode: StoredPermissionMode;
-}
-
-interface EffectiveCreateExecutionValues {
-  selectedProviderId: string;
-  selectedModel: string;
-  serviceTier: ServiceTier | undefined;
-  reasoningLevel: ReasoningLevel;
-  permissionMode: PermissionMode;
-}
-
-interface BuildExecutionInputSourcesArgs {
-  effectiveValues: EffectiveCreateExecutionValues;
-  scope: ThreadCreationOptionsScope;
-  storedValues: StoredCreateExecutionValues;
-  touchedFields: ReadonlySet<ThreadPromptField>;
-}
-
-interface SyncThreadPromptSelectionsArgs {
-  currentSelections: ThreadPromptSelections;
-  nextSelections: ThreadPromptSelections;
-  touchedFields: ReadonlySet<ThreadPromptField>;
-}
-
-interface UpdateThreadPromptSelectionsArgs {
-  currentSelections: ThreadPromptSelections;
-  field: ThreadPromptField;
-  value: ThreadPromptSelections[ThreadPromptField];
-}
-
-interface ResolvePermissionModeSelectionArgs {
-  rawPermissionMode: PermissionMode;
-  supportedPermissionModes: readonly PermissionMode[];
-}
-
 const NO_MODEL_LOAD_ERROR: SystemExecutionOptionsModelLoadError | null = null;
-
-function isReasoningLevel(value: unknown): value is ReasoningLevel {
-  return (
-    value === "low" ||
-    value === "medium" ||
-    value === "high" ||
-    value === "xhigh" ||
-    value === "max"
-  );
-}
-
-function isPermissionMode(value: unknown): value is PermissionMode {
-  return (
-    value === "readonly" || value === "workspace-write" || value === "full"
-  );
-}
-
-function isServiceTier(value: unknown): value is ServiceTier {
-  return value === "fast" || value === "default";
-}
-
-function isStoredServiceTier(value: string): value is StoredServiceTier {
-  return value === "" || isServiceTier(value);
-}
-
-function isStoredReasoningLevel(value: string): value is StoredReasoningLevel {
-  return value === "" || isReasoningLevel(value);
-}
-
-function isStoredPermissionMode(value: string): value is StoredPermissionMode {
-  return value === "" || isPermissionMode(value);
-}
-
-const storedServiceTierStorage =
-  createLocalStorageEnumStorage<StoredServiceTier>(isStoredServiceTier);
-const storedReasoningLevelStorage =
-  createLocalStorageEnumStorage<StoredReasoningLevel>(isStoredReasoningLevel);
-const storedPermissionModeStorage =
-  createLocalStorageEnumStorage<StoredPermissionMode>(isStoredPermissionMode);
-const providerIdAtomFamily = createProjectScopedStorageAtomFamily(
-  PROVIDER_STORAGE_KEY,
-  "",
-  rawStringLocalStorage,
-);
-const modelAtomFamily = createProjectScopedStorageAtomFamily(
-  MODEL_STORAGE_KEY,
-  "",
-  rawStringLocalStorage,
-);
-const serviceTierAtomFamily =
-  createProjectScopedStorageAtomFamily<StoredServiceTier>(
-    SERVICE_TIER_STORAGE_KEY,
-    "",
-    storedServiceTierStorage,
-  );
-const reasoningLevelAtomFamily = createProjectScopedStorageAtomFamily(
-  REASONING_STORAGE_KEY,
-  "",
-  storedReasoningLevelStorage,
-);
-const permissionModeAtomFamily = createProjectScopedStorageAtomFamily(
-  PERMISSION_MODE_STORAGE_KEY,
-  "",
-  storedPermissionModeStorage,
-);
-const environmentSelectionAtomFamily = createProjectScopedStorageAtomFamily(
-  ENVIRONMENT_STORAGE_KEY,
-  "",
-  rawStringLocalStorage,
-);
-
-function getInitialThreadPromptSelections(
-  options?: UsePromptModelReasoningOptions,
-): ThreadPromptSelections {
-  return {
-    selectedProviderId: options?.initialProviderId ?? "",
-    selectedModel: options?.initialModel ?? "",
-    serviceTier: options?.initialServiceTier,
-    reasoningLevel: options?.initialReasoningLevel ?? "medium",
-    permissionMode: options?.initialPermissionMode ?? "full",
-    environmentSelectionValue: options?.initialEnvironmentSelectionValue ?? "",
-  };
-}
-
-function syncUntouchedThreadPromptSelections({
-  currentSelections,
-  nextSelections,
-  touchedFields,
-}: SyncThreadPromptSelectionsArgs): ThreadPromptSelections {
-  let changed = false;
-  const updatedSelections = { ...currentSelections };
-
-  if (
-    !touchedFields.has("selectedProviderId") &&
-    currentSelections.selectedProviderId !== nextSelections.selectedProviderId
-  ) {
-    updatedSelections.selectedProviderId = nextSelections.selectedProviderId;
-    changed = true;
-  }
-  if (
-    !touchedFields.has("selectedModel") &&
-    currentSelections.selectedModel !== nextSelections.selectedModel
-  ) {
-    updatedSelections.selectedModel = nextSelections.selectedModel;
-    changed = true;
-  }
-  if (
-    !touchedFields.has("serviceTier") &&
-    currentSelections.serviceTier !== nextSelections.serviceTier
-  ) {
-    updatedSelections.serviceTier = nextSelections.serviceTier;
-    changed = true;
-  }
-  if (
-    !touchedFields.has("reasoningLevel") &&
-    currentSelections.reasoningLevel !== nextSelections.reasoningLevel
-  ) {
-    updatedSelections.reasoningLevel = nextSelections.reasoningLevel;
-    changed = true;
-  }
-  if (
-    !touchedFields.has("permissionMode") &&
-    currentSelections.permissionMode !== nextSelections.permissionMode
-  ) {
-    updatedSelections.permissionMode = nextSelections.permissionMode;
-    changed = true;
-  }
-  if (
-    !touchedFields.has("environmentSelectionValue") &&
-    currentSelections.environmentSelectionValue !==
-      nextSelections.environmentSelectionValue
-  ) {
-    updatedSelections.environmentSelectionValue =
-      nextSelections.environmentSelectionValue;
-    changed = true;
-  }
-
-  return changed ? updatedSelections : currentSelections;
-}
-
-function updateThreadPromptSelections({
-  currentSelections,
-  field,
-  value,
-}: UpdateThreadPromptSelectionsArgs): ThreadPromptSelections {
-  if (currentSelections[field] === value) {
-    return currentSelections;
-  }
-
-  return {
-    ...currentSelections,
-    [field]: value,
-  };
-}
-
-function hasValue(value: string): boolean {
-  return value.length > 0;
-}
-
-function resolveCreateExecutionInputSource({
-  hasStoredValue,
-  hasValue,
-  touched,
-}: ResolveCreateExecutionInputSourceArgs): ExecutionInputFieldSource | undefined {
-  if (!hasValue) {
-    return undefined;
-  }
-  if (touched) {
-    return "explicit";
-  }
-  if (hasStoredValue) {
-    return "client-preference";
-  }
-  return undefined;
-}
-
-function buildExecutionInputSources({
-  effectiveValues,
-  scope,
-  storedValues,
-  touchedFields,
-}: BuildExecutionInputSourcesArgs): ScopedExecutionInputSources {
-  const usesStoredValues = scope === "new-thread";
-  const hasTouchedExecutionField =
-    touchedFields.has("selectedProviderId") ||
-    touchedFields.has("selectedModel") ||
-    touchedFields.has("serviceTier") ||
-    touchedFields.has("reasoningLevel") ||
-    touchedFields.has("permissionMode");
-  // Existing-thread submissions are all-or-nothing once an execution control is
-  // touched, so the server never merges stale last-run values with new UI picks.
-  const forcesExplicitExecutionFields =
-    scope === "component-local" && hasTouchedExecutionField;
-
-  if (!usesStoredValues && scope !== "component-local") {
-    return {};
-  }
-
-  const providerSource = resolveCreateExecutionInputSource({
-    hasStoredValue:
-      usesStoredValues &&
-      hasValue(storedValues.selectedProviderId) &&
-      storedValues.selectedProviderId === effectiveValues.selectedProviderId,
-    hasValue: hasValue(effectiveValues.selectedProviderId),
-    touched: touchedFields.has("selectedProviderId"),
-  });
-  const modelSource = resolveCreateExecutionInputSource({
-    hasStoredValue:
-      usesStoredValues &&
-      hasValue(storedValues.selectedModel) &&
-      storedValues.selectedModel === effectiveValues.selectedModel,
-    hasValue: hasValue(effectiveValues.selectedModel),
-    touched:
-      forcesExplicitExecutionFields || touchedFields.has("selectedModel"),
-  });
-  const serviceTierSource = resolveCreateExecutionInputSource({
-    hasStoredValue:
-      usesStoredValues &&
-      storedValues.serviceTier !== "" &&
-      storedValues.serviceTier === effectiveValues.serviceTier,
-    hasValue: effectiveValues.serviceTier !== undefined,
-    touched:
-      forcesExplicitExecutionFields || touchedFields.has("serviceTier"),
-  });
-  const reasoningLevelSource = resolveCreateExecutionInputSource({
-    hasStoredValue: usesStoredValues && storedValues.reasoningLevel !== "",
-    hasValue: hasValue(effectiveValues.reasoningLevel),
-    touched:
-      forcesExplicitExecutionFields || touchedFields.has("reasoningLevel"),
-  });
-  const permissionModeSource = resolveCreateExecutionInputSource({
-    hasStoredValue: usesStoredValues && storedValues.permissionMode !== "",
-    hasValue: hasValue(effectiveValues.permissionMode),
-    touched:
-      forcesExplicitExecutionFields || touchedFields.has("permissionMode"),
-  });
-
-  if (scope === "component-local") {
-    return {
-      ...(modelSource ? { model: modelSource } : {}),
-      ...(serviceTierSource ? { serviceTier: serviceTierSource } : {}),
-      ...(reasoningLevelSource ? { reasoningLevel: reasoningLevelSource } : {}),
-      ...(permissionModeSource
-        ? { permissionMode: permissionModeSource }
-        : {}),
-    };
-  }
-
-  return {
-    ...(providerSource ? { providerId: providerSource } : {}),
-    ...(modelSource ? { model: modelSource } : {}),
-    ...(serviceTierSource ? { serviceTier: serviceTierSource } : {}),
-    ...(reasoningLevelSource ? { reasoningLevel: reasoningLevelSource } : {}),
-    ...(permissionModeSource
-      ? { permissionMode: permissionModeSource }
-      : {}),
-  };
-}
-
-export function resolvePermissionModeSelection({
-  rawPermissionMode,
-  supportedPermissionModes,
-}: ResolvePermissionModeSelectionArgs): PermissionMode {
-  if (supportedPermissionModes.includes(rawPermissionMode)) {
-    return rawPermissionMode;
-  }
-  if (supportedPermissionModes.includes("full")) {
-    return "full";
-  }
-  return supportedPermissionModes[0] ?? "full";
-}
-
-export function formatModelLabel(value: string): string {
-  // Case-normalises a raw model id into a displayable label. The brand prefix
-  // strip ("Claude " / "GPT-") is a presentation rule applied by the picker
-  // itself (see `stripModelBrandPrefix`) so stories and prod render identically
-  // without anyone having to remember to format.
-  return value
-    .split("-")
-    .map((part) => {
-      if (part.toLowerCase() === "gpt") return "GPT";
-      if (/^\d+(\.\d+)*$/.test(part)) return part;
-      if (/^[a-z]+$/i.test(part)) {
-        return part.charAt(0).toUpperCase() + part.slice(1).toLowerCase();
-      }
-      return part;
-    })
-    .join("-");
-}
 
 function sanitizeStoredEnvironmentValue(stored: string): string {
   // Legacy guard: earlier iterations briefly persisted `reuse:<envId>` to
@@ -522,23 +153,20 @@ export function useThreadCreationOptions(
     resetKey,
     scope = "new-thread",
   } = options ?? {};
-  const [storedProviderId, setStoredProviderId] = useAtom(
-    providerIdAtomFamily(projectId),
-  );
-  const [storedSelectedModel, setStoredSelectedModel] = useAtom(
-    modelAtomFamily(projectId),
-  );
-  const [storedServiceTier, setStoredServiceTier] = useAtom(
-    serviceTierAtomFamily(projectId),
-  );
-  const [storedReasoningLevel, setStoredReasoningLevel] = useAtom(
-    reasoningLevelAtomFamily(projectId),
-  );
-  const [storedPermissionMode, setStoredPermissionMode] = useAtom(
-    permissionModeAtomFamily(projectId),
-  );
-  const [storedEnvironmentSelectionValue, setStoredEnvironmentSelectionValue] =
-    useAtom(environmentSelectionAtomFamily(projectId));
+  const { setValue: setStoredProviderId, value: storedProviderId } =
+    usePersistedProviderSelection(projectId);
+  const { setValue: setStoredSelectedModel, value: storedSelectedModel } =
+    usePersistedModelSelection(projectId);
+  const { setValue: setStoredServiceTier, value: storedServiceTier } =
+    usePersistedServiceTierSelection(projectId);
+  const { setValue: setStoredReasoningLevel, value: storedReasoningLevel } =
+    usePersistedReasoningLevelSelection(projectId);
+  const { setValue: setStoredPermissionMode, value: storedPermissionMode } =
+    usePersistedPermissionModeSelection(projectId);
+  const {
+    setValue: setStoredEnvironmentSelectionValue,
+    value: storedEnvironmentSelectionValue,
+  } = usePersistedEnvironmentSelection(projectId);
   // Reuse env values are intentionally NEVER persisted to localStorage —
   // they represent a transient "create one thread in this worktree" intent,
   // not a project default.
@@ -603,26 +231,21 @@ export function useThreadCreationOptions(
     usesLocalThreadSelections,
   ]);
 
-  const rawSelectedProviderId =
-    usesStoredCreateSelections
-      ? storedProviderId || renderedThreadSelections.selectedProviderId
-      : renderedThreadSelections.selectedProviderId;
-  const rawSelectedModel =
-    usesStoredCreateSelections
-      ? storedSelectedModel || renderedThreadSelections.selectedModel
-      : renderedThreadSelections.selectedModel;
-  const rawServiceTier =
-    usesStoredCreateSelections
-      ? storedServiceTier || renderedThreadSelections.serviceTier
-      : renderedThreadSelections.serviceTier;
-  const rawReasoningLevel =
-    usesStoredCreateSelections
-      ? storedReasoningLevel || renderedThreadSelections.reasoningLevel
-      : renderedThreadSelections.reasoningLevel;
-  const rawPermissionMode =
-    usesStoredCreateSelections
-      ? storedPermissionMode || renderedThreadSelections.permissionMode
-      : renderedThreadSelections.permissionMode;
+  const rawSelectedProviderId = usesStoredCreateSelections
+    ? storedProviderId || renderedThreadSelections.selectedProviderId
+    : renderedThreadSelections.selectedProviderId;
+  const rawSelectedModel = usesStoredCreateSelections
+    ? storedSelectedModel || renderedThreadSelections.selectedModel
+    : renderedThreadSelections.selectedModel;
+  const rawServiceTier = usesStoredCreateSelections
+    ? storedServiceTier || renderedThreadSelections.serviceTier
+    : renderedThreadSelections.serviceTier;
+  const rawReasoningLevel = usesStoredCreateSelections
+    ? storedReasoningLevel || renderedThreadSelections.reasoningLevel
+    : renderedThreadSelections.reasoningLevel;
+  const rawPermissionMode = usesStoredCreateSelections
+    ? storedPermissionMode || renderedThreadSelections.permissionMode
+    : renderedThreadSelections.permissionMode;
   const rawEnvironmentSelectionValue =
     scope === "new-thread"
       ? (rootComposeReuseValue ??
